requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=pyhd3eb1b0_0
appdirs=1.4.4=py_0
argh=0.26.2=py39haa95532_0
astroid=2.5=py39haa95532_1
async_generator=1.10=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.6=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py39h196d8e1_0
black=19.3b0=py_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2020.12.5=h5b45459_0
certifi=2020.12.5=py39hcbf5309_1
cffi=1.14.5=py39hcd4344a_0
chardet=4.0.0=py39haa95532_1003
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.7=py39h71e12ea_0
cycler=0.10.0=py39haa95532_0
cython=0.29.22=py39hd77b12b_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=py_0
docutils=0.16=py39haa95532_1
entrypoints=0.3=py39haa95532_0
flake8=3.9.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39haa95532_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imagesize=1.2.0=pyhd3eb1b0_0
importlib-metadata=3.7.3=py39haa95532_1
importlib_metadata=3.7.3=hd3eb1b0_1
intel-openmp=2020.2=254
intervaltree=3.1.0=py_0
ipykernel=5.3.4=py39h7b7c402_0
ipython=7.21.0=py39hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.8.0=pyhd3eb1b0_0
jedi=0.17.2=py39haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
keyring=22.3.0=py39haa95532_0
kiwisolver=1.3.1=py39hd77b12b_0
lazy-object-proxy=1.6.0=py39h2bbff1b_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.1.0=h56a325e_1
littleutils=0.2.2=py_0
lz4-c=1.9.3=h2bbff1b_0
markupsafe=1.1.1=py39h2bbff1b_0
matplotlib=3.3.4=py39haa95532_0
matplotlib-base=3.3.4=py39h49ac443_0
mccabe=0.6.1=py39haa95532_1
mistune=0.8.4=py39h2bbff1b_1000
mkl=2020.2=256
mkl-service=2.3.0=py39h196d8e1_0
mkl_fft=1.3.0=py39h46781fe_0
mkl_random=1.0.2=py39h848d8c7_0
mpmath=1.2.1=pyhd8ed1ab_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py39haa95532_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
numpy=1.19.2=py39h729668d_0
numpy-base=1.19.2=py39hbd0edd7_0
numpydoc=1.1.0=pyhd3eb1b0_1
olefile=0.46=py_0
openssl=1.1.1k=h8ffe710_0
outdated=0.2.0=py_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.3=py39hf11a4ad_0
pandas-flavor=0.2.0=py_0
pandoc=2.12=haa95532_0
pandocfilters=1.4.3=py39haa95532_1
paramiko=2.7.2=py_0
parso=0.7.0=py_0
patsy=0.5.1=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py39h4fa10fc_0
pingouin=0.3.10=pyhd8ed1ab_0
pip=21.0.1=py39haa95532_0
pluggy=0.13.1=py39haa95532_0
prompt-toolkit=3.0.17=pyh06a4308_0
psutil=5.8.0=py39h2bbff1b_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pycodestyle=2.6.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydocstyle=6.0.0=pyhd3eb1b0_0
pyflakes=2.2.0=pyhd3eb1b0_0
pygame=2.0.1=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pylint=2.7.2=py39haa95532_1
pyls-black=0.4.6=hd3eb1b0_0
pyls-spyder=0.3.2=pyhd3eb1b0_0
pynacl=1.4.0=py39hbd8134f_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.17.3=py39h2bbff1b_0
pysocks=1.7.1=py39haa95532_0
python=3.9.2=h6244533_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-jsonrpc-server=0.4.0=py_0
python-language-server=0.36.2=pyhd3eb1b0_0
python-pygaze=0.7.1=pypi_0
python_abi=3.9=1_cp39
pytz=2021.1=pyhd3eb1b0_0
pywin32=228=py39he774522_0
pywin32-ctypes=0.2.0=py39haa95532_1000
pyyaml=5.4.1=py39h2bbff1b_1
pyzmq=20.0.0=py39hd77b12b_1
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.2=pyhd3eb1b0_0
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
requests=2.25.1=pyhd3eb1b0_0
rope=0.18.0=py_0
rtree=0.9.4=py39h2eaa2aa_1
scikit-learn=0.24.1=py39hf11a4ad_0
scipy=1.6.2=py39h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py39haa95532_0
sip=4.19.13=py39hd77b12b_0
six=1.15.0=py39haa95532_0
snowballstemmer=2.1.0=pyhd3eb1b0_0
sortedcontainers=2.3.0=pyhd3eb1b0_0
sphinx=3.5.3=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.4=pyhd3eb1b0_0
spyder=4.2.5=py39haa95532_0
spyder-kernels=1.10.2=py39haa95532_0
sqlite=3.35.2=h2bbff1b_0
statsmodels=0.12.2=py39h2bbff1b_0
tabulate=0.8.9=pyhd8ed1ab_0
testpath=0.4.4=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
three-merge=0.1.1=pyhd3eb1b0_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py39h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
tzdata=2020f=h52ac0ba_0
ujson=4.0.2=py39hd77b12b_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=1.0.2=py39haa95532_1
wcwidth=0.2.5=py_0
webencodings=0.5.1=py39haa95532_1
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39h2bbff1b_0
wrapt=1.12.1=py39h196d8e1_1
xarray=0.17.0=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
